+++ AVR3560Test/Steppers/xStepper.h
@@ -8,8 +8,8 @@
 #define XSTEPPER_H_
 
 struct Stepper{
-	 volatile unsigned int steps_to_do;
-	 volatile unsigned int done_steps;
+	 volatile unsigned long steps_to_do;
+	 volatile unsigned long done_steps;
 };
 
 void enableX();
